Revert "Update to 3b470f56b479f618b7e90577f9857b4e25b38c1a"
This reverts commit 04f21656a55bdb0cc0212fdb5439cd3ca900785a.
This was done wrongly on master. Should be done on upstream branch

+++ core/src/main/scala/kafka/javaapi/message/MessageSet.scala
@@ -38,7 +38,7 @@ abstract class MessageSet extends java.lang.Iterable[MessageAndOffset] {
   /**
    * Gives the total size of this message set in bytes
    */
-  def sizeInBytes: Int
+  def sizeInBytes: Long
 
   /**
    * Validate the checksum of all the messages in the set. Throws an InvalidMessageException if the checksum doesn't
