+++ core/src/test/scala/unit/kafka/javaapi/message/BaseMessageSetTestCases.scala
@@ -58,17 +58,17 @@ trait BaseMessageSetTestCases extends JUnitSuite {
   @Test
   def testSizeInBytes() {
     assertEquals("Empty message set should have 0 bytes.",
-                 0,
+                 0L,
                  createMessageSet(Array[Message]()).sizeInBytes)
     assertEquals("Predicted size should equal actual size.", 
-                 kafka.message.MessageSet.messageSetSize(messages),
+                 kafka.message.MessageSet.messageSetSize(messages).toLong,
                  createMessageSet(messages).sizeInBytes)
   }
 
   @Test
   def testSizeInBytesWithCompression () {
     assertEquals("Empty message set should have 0 bytes.",
-                 0,           // overhead of the GZIP output stream
+                 0L,           // overhead of the GZIP output stream
                  createMessageSet(Array[Message](), DefaultCompressionCodec).sizeInBytes)
   }
 }
